Began work on HTTP parsing following RFC 2616
HTTP headers are parsed following a state machine
when data is encountered, it will be inserted into
an http object.

To this point, the parser only successfully parses the HTTP Version
string from the data, or checks to see if it is a request instead.
A testing file has been added, though it does not test much as of now.